% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahull3D.R
\name{ahull3D}
\alias{ahull3D}
\title{Fast 3D Alpha Hull with Label Propagation}
\usage{
ahull3D(points, alpha, input_truth = NULL, volume = FALSE)
}
\arguments{
\item{points}{Nx3 matrix of points}

\item{alpha}{Alpha value}

\item{input_truth}{Optional labels (point id)}

\item{volume}{Compute volume (default: FALSE)}
}
\description{
Fast 3D Alpha Hull with Label Propagation
}
