% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahull3D.R
\name{ahull3d}
\alias{ahull3d}
\title{Fast 3D Alpha Hull with Label Propagation}
\usage{
ahull3d(points, alpha, input_truth = NULL, volume = FALSE)
}
\arguments{
\item{points}{Nx3 matrix of points}

\item{alpha}{Alpha value}

\item{input_truth}{Optional labels}
}
\description{
Fast 3D Alpha Hull with Label Propagation
}
